# lib/api-zod/src/generated/types/regionSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegionSnapshot {
  name: string;
  stationCount: number;
  cheapestPetrol: number;
  cheapestDiesel: number;
  cheapestEv: number;
}